requirements.txt
python-telegram-bot==22.1
python-dotenv==1.1.0
structlog==25.4.0
pydantic==2.11.5
pydantic-settings==2.9.1